+++ src/main/java/net/kccricket/bestesttool/commands/CommandReload.java
@@ -2,7 +2,6 @@ package net.kccricket.bestesttool.commands;
 
 import net.kccricket.bestesttool.BestestToolPlugin;
 import net.kccricket.bestesttool.security.Permissions;
-import net.kccricket.bestesttool.text.MessageUtil;
 
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
 
@@ -13,11 +12,11 @@ public class CommandReload {
     static void reload(CommandSender sender, BestestToolPlugin main) {
 
             if (!Permissions.isAllowedTo(sender, Permissions.PERM_RELOAD)) {
-                MessageUtil.send(sender, "noPermission", Placeholder.unparsed("plugin", main.getName()));
+                main.messages().to(sender).error().send("noPermission", Placeholder.unparsed("plugin", main.getName()));
                 return;
             }
             main.reload();
-            MessageUtil.send(sender, "reloaded", Placeholder.unparsed("plugin", main.getName()));
+            main.messages().to(sender).status().send("reloaded", Placeholder.unparsed("plugin", main.getName()));
     }
 
 }